pyproject.toml
[tool.poetry]
name = "visitor-backend"
version = "0.1.0"
description = ""
authors = ["Priyosmita Nandy <priyosh08@gmail.com>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
odmantic = "^1.0.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
isort = "^5.13.2"
black = "^23.12.1"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.2"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
